Changed clue entry to scale widgets
This way, you don't have to use both keyboard and mouse.  This allowed
cleanup of a lot of code, since user can't entehr bad clues.  I also
disabled the scales in squares where clues can't be enetered.  I changed
the CSP solver to limit the domains of the variables to possibilities
depending on the clues, instead of always making them 1-9, but I haven't
done any testing.